requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
matplotlib
seaborn
streamlit
pandas
scikit-learn
joblib
tensorflow==2.12.0
pillow
opencv-python
kaggle
matplotlib
